pyproject.toml
[project]
name = "telaten-backend"
version = "0.1.0"
description = "Backend for Telaten - AI Powered MSME Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.109.0",
  "sqlalchemy>=2.0.25",
  "pydantic-settings>=2.1.0",
  "uvicorn>=0.38.0",
  "sqlmodel>=0.0.27",
  "asyncpg>=0.31.0",
  "passlib>=1.7.4",
  "python-jose[cryptography]>=3.5.0",
  "structlog>=25.5.0",
  "llama-index>=0.14.8",
  "llama-index-llms-openai-like>=0.5.3",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
